Removed the revisionist research questions
I introduced these questions retrospectively, when I started writing, as a way to try to frame this paper. They didn't really come up in the findings, though, and when I went to write the conclusion, the pieces were not fitting together. I think this is because the project was really wrapped up in the design goals, not the research questions.

Future drafts might delve a bit more into the nature of design-based research, but then again, the readership of this piece may not care about that.